Added some moethods for BayesMallowsDPMixture objects and bug fixes
Added argument clus_thin in tidy_cluster_assignment, will be needed when setting burnin for computation of co_clustering matrix

+++ src/dpmixtures.cpp
@@ -10,7 +10,7 @@ using namespace arma;
 // [[Rcpp::depends(RcppArmadillo)]]
 
 
-vec initial_assignment_prob(const uvec already_assigned, const int psi, const int n_clus){
+vec initial_assignment_prob(const uvec already_assigned, const double psi, const int n_clus){
     vec prob(n_clus + 1);
     for(int i = 0; i < n_clus; ++i){
         uvec elem_i = find(already_assigned == i);
@@ -20,7 +20,7 @@ vec initial_assignment_prob(const uvec already_assigned, const int psi, const in
     return normalise(prob, 1);
 }
 
-uvec initialize_cluster_assignment(const int n_assessors, const int psi){
+uvec initialize_cluster_assignment(const int n_assessors, const double psi){
     uvec cluster_assignment(n_assessors);
     cluster_assignment(0) = 0;
     int n_clus = 1;
@@ -67,7 +67,7 @@ uvec update_cluster_labels_dpmixture(const mat& rankings,
                                     const double& log_fact_n_items,
                                     const double& lambda,
                                     const double& alpha_max,
-                                    const int& psi,
+                                    const double& psi,
                                     const int& leap_size,
                                     const std::string& metric,
                                     const Rcpp::Nullable<vec> cardinalities = R_NilValue,
